requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework==3.12.4
django-extensions==3.1.3
django-model-utils==4.0.0
